Refactor to make printing objects more type-safe
