+++ packages/jumpstarter-cli-common/jumpstarter_cli_common/__init__.py
@@ -1,5 +1,5 @@
 from .alias import AliasedGroup
-from .opt import opt_context, opt_kubeconfig, opt_labels, opt_log_level, opt_namespace, opt_output
+from .opt import OutputType, opt_context, opt_kubeconfig, opt_labels, opt_log_level, opt_namespace, opt_output
 from .table import make_table
 from .time import time_since
 from .version import get_client_version, version
@@ -13,6 +13,7 @@ __all__ = [
     "opt_namespace",
     "opt_labels",
     "opt_output",
+    "OutputType",
     "time_since",
     "version",
     "get_client_version",
